+++ Core/src/com/mfino/domain/PocketTemplateConfig.java
@@ -45,8 +45,8 @@ public class PocketTemplateConfig extends Base implements java.io.Serializable {
 	
 	private KycLevel kycLevel;
 	private PocketTemplate pocketTemplate;
-	private long subscribertype;
-	private Long businesspartnertype;
+	private int subscribertype;
+	private Integer businesspartnertype;
 	private long commodity;
 	private long pockettype;
 	private Short issuspencepocket;
@@ -94,20 +94,20 @@ public class PocketTemplateConfig extends Base implements java.io.Serializable {
 	
 
 	@Column(name = "SUBSCRIBERTYPE", nullable = false, precision = 10, scale = 0)
-	public long getSubscribertype() {
+	public int getSubscribertype() {
 		return this.subscribertype;
 	}
 
-	public void setSubscribertype(long subscribertype) {
+	public void setSubscribertype(int subscribertype) {
 		this.subscribertype = subscribertype;
 	}
 
 	@Column(name = "BUSINESSPARTNERTYPE", precision = 10, scale = 0)
-	public Long getBusinesspartnertype() {
+	public Integer getBusinesspartnertype() {
 		return this.businesspartnertype;
 	}
 
-	public void setBusinesspartnertype(Long businesspartnertype) {
+	public void setBusinesspartnertype(Integer businesspartnertype) {
 		this.businesspartnertype = businesspartnertype;
 	}
